+++ src/GridCal/Engine/Simulations/result_types.py
@@ -26,6 +26,7 @@ class ResultTypes(Enum):
     BranchPower = 'Branch power', DeviceType.BranchDevice
     BranchCurrent = 'Branch current', DeviceType.BranchDevice
     BranchLoading = 'Branch loading', DeviceType.BranchDevice
+    BranchTapModule = 'Branch tap module', DeviceType.BranchDevice
     BranchVoltage = 'Branch voltage drop', DeviceType.BranchDevice
     BranchAngles = 'Branch voltage angles', DeviceType.BranchDevice
     BranchLosses = 'Branch losses', DeviceType.BranchDevice